(Module 09 ex00) Properly turn my map into a string, flaot instead of a string, double

+++ Module_09/ex00/incs/header.hpp
@@ -10,7 +10,7 @@
 #include <sstream>
 #include <algorithm>
 
-typedef std::map<std::string, double> myMap;
+typedef std::map<std::string, float> myMap;
 
 bool	validDateFormat(const std::string& date);
 bool	invalidDate(const std::string& date, myMap map);
